[JSC] Report extra memory allocated by JSC::Exception and JSC::JSWebAssemblyException https://bugs.webkit.org/show_bug.cgi?id=283108
Reviewed by NOBODY (OOPS!).

JSTests/wasm/stress/cc-int-to-int-cross-module-with-exception.js would
exceed the memory footprint limit when run with --memory-limited. This
was due to it allocating a lot of exceptions, while the size of the
allocated exceptions reported to the GC was wildly underestimated.
This lead to the GC kicking in too late because it is unable to track the
true eden size.

This is resolved by calling Heap::reportExtraMemoryAllocated on
on the stacktrace and payload allocated by JSC::Exception and
JSC::JSWebAssemblyException respectively.

* Source/JavaScriptCore/runtime/Exception.cpp:
(JSC::Exception::finishCreation):
* Source/JavaScriptCore/wasm/js/JSWebAssemblyException.cpp:
(JSC::JSWebAssemblyException::finishCreation):
(JSC::JSWebAssemblyException::visitChildrenImpl):
* Source/JavaScriptCore/wasm/js/JSWebAssemblyException.h: